EZP-26885: Moved Legacy StorageHandler to common namespace
